add project enum import.

+++ DevCopilot2.Web/PresentationExtensions/IdentityExtensions.cs
@@ -5,34 +5,34 @@ namespace DevCopilot2.Web.PresentationExtensions
 {
     public static class IdentityExtensions
     {
-        public static long GetCurrentUserId(this ClaimsPrincipal claimsPrincipal)
+        public static int GetCurrentUserId(this ClaimsPrincipal claimsPrincipal)
         {
             if (claimsPrincipal != null)
             {
                 var data = claimsPrincipal.Claims.SingleOrDefault(s => s.Type == ClaimTypes.NameIdentifier);
-                if (data != null) return Convert.ToInt64(data.Value);
+                if (data != null) return Convert.ToInt32(data.Value);
             }
 
-            return default(long);
+            return default(int);
             //return 3;
         }
-        public static long? GetNullableCurrentUserId(this ClaimsPrincipal claimsPrincipal)
+        public static int? GetNullableCurrentUserId(this ClaimsPrincipal claimsPrincipal)
         {
             if (claimsPrincipal != null)
             {
                 var data = claimsPrincipal.Claims.SingleOrDefault(s => s.Type == ClaimTypes.NameIdentifier);
-                if (data != null) return Convert.ToInt64(data.Value);
+                if (data != null) return Convert.ToInt32(data.Value);
             }
             return null;
         }
 
-        public static long GetCurrentUserId(this IPrincipal principal)
+        public static int GetCurrentUserId(this IPrincipal principal)
         {
             var user = (ClaimsPrincipal)principal;
             return user.GetCurrentUserId();
         }
 
-        public static long? GetNullableCurrentUserId(this IPrincipal principal)
+        public static int? GetNullableCurrentUserId(this IPrincipal principal)
         {
             var user = (ClaimsPrincipal)principal;
             return user.GetNullableCurrentUserId();
